+++ src/arpes/utilities/conversion/kz_conversion.py
@@ -20,7 +20,6 @@ if TYPE_CHECKING:
     from _typeshed import Incomplete
     from numpy.typing import NDArray
 
-    from arpes._typing import MOMENTUM
 
 __all__ = ["ConvertKpKz", "ConvertKpKzV0", "ConvertKxKyKz"]
 
@@ -86,8 +85,8 @@ class ConvertKpKz(CoordinateConverter):
 
     def get_coordinates(
         self,
-        resolution: dict[MOMENTUM, float] | None = None,
-        bounds: dict[MOMENTUM, tuple[float, float]] | None = None,
+        resolution: dict[str, float] | None = None,
+        bounds: dict[str, tuple[float, float]] | None = None,
     ) -> dict[Hashable, NDArray[np.float64]]:
         """Calculates the coordinates which should be used in momentum space.
 
@@ -103,7 +102,9 @@ class ConvertKpKz(CoordinateConverter):
         resolution = resolution if resolution is not None else {}
         assert resolution is not None
         bounds = bounds if bounds is not None else {}
-        coordinates = super().get_coordinates(resolution=resolution, bounds=bounds)
+
+        coordinates = {k: v.values for k, v in self.arr.coords.items() if k == "eV"}
+
         ((kp_low, kp_high), (kz_low, kz_high)) = calculate_kp_kz_bounds(self.arr)
         if "kp" in bounds:
             kp_low, kp_high = bounds["kp"]
@@ -119,11 +120,13 @@ class ConvertKpKz(CoordinateConverter):
             kp_low - K_SPACE_BORDER,
             kp_high + K_SPACE_BORDER,
             resolution.get("kp", inferred_kp_res),
+            dtype=np.float64,
         )
         coordinates["kz"] = np.arange(
             kz_low - K_SPACE_BORDER,
             kz_high + K_SPACE_BORDER,
             resolution.get("kz", inferred_kz_res),
+            dtype=np.float64,
         )
         base_coords = {
             k: v.values for k, v in self.arr.coords.items() if k not in {"eV", "phi", "hv"}