+++ tests/test_config.py
@@ -1,14 +1,13 @@
-from pathlib import Path
-from unittest.mock import patch
-
 import pytest
-
+from unittest.mock import patch, MagicMock
+from pathlib import Path
+import logging
 import arpes.config
 
 
 @pytest.fixture
 def mock_cwd():
-    with patch("arpes.config.Path.cwd") as mock:
+    with patch("pathlib.Path.cwd") as mock:
         yield mock
 
 
@@ -20,7 +19,7 @@ def mock_workspace_matches():
 
 @pytest.fixture
 def mock_logging_exception():
-    with patch("arpes.config.logging.exception") as mock:
+    with patch("logging.exception") as mock:
         yield mock
 
 
@@ -41,18 +40,5 @@ def test_attempt_determine_workspace_not_found(mock_cwd, mock_workspace_matches)
     arpes.config.DATASET_PATH = Path("/mock/dataset")
     arpes.config.attempt_determine_workspace()
 
-    assert arpes.config.CONFIG["WORKSPACE"]["path"] == Path("/mock/dataset")
-    assert arpes.config.CONFIG["WORKSPACE"]["name"] == "dataset"
-
-
-def test_attempt_determine_workspace_exception(
-    mock_cwd, mock_workspace_matches, mock_logging_exception,
-):
-    mock_cwd.side_effect = Exception("Test Exception")
-
-    arpes.config.DATASET_PATH = Path("/mock/dataset")
-    arpes.config.attempt_determine_workspace()
-
-    mock_logging_exception.assert_called_once_with("Exception occurs")
     assert arpes.config.CONFIG["WORKSPACE"]["path"] == Path("/mock/dataset")
     assert arpes.config.CONFIG["WORKSPACE"]["name"] == "dataset"